// src/domain/types/v1api/resUploadedV2.ts
/**
 * Generated by orval. Do not edit manually.
 * O-replay - OpenAPI 3.0
 * O-replay Rest API
 */
import type { ResUploadedV2Data } from "./resUploadedV2Data"
import type { UploadedV2Meta } from "./uploadedV2Meta"

/**
 * Data wrapper for ResUploadedV2
 */
export interface ResUploadedV2 {
  /** Any object */
  data: ResUploadedV2Data
  meta?: UploadedV2Meta
}