// src/domain/types/v1api/uploadMessage.ts
/**
 * Generated by orval. Do not edit manually.
 * O-replay - OpenAPI 3.0
 * O-replay Rest API
 */
import type { UploadMessageContext } from "./uploadMessageContext"

/**
 * Entity UploadMessage
 */
export interface UploadMessage {
  code?: string
  /** Generic object. */
  context?: UploadMessageContext
  level?: string
  text?: string
}